+++ zython/_compile/zinc/flags.py
@@ -9,6 +9,7 @@ class Flags(enum.Flag):
     none = enum.auto()
     alldifferent = enum.auto()
     alldifferent_except_0 = enum.auto()
+    alldifferent_except = enum.auto()
     all_equal = enum.auto()
     nvalue = enum.auto()
     circuit = enum.auto()
@@ -17,12 +18,14 @@ class Flags(enum.Flag):
     decreasing = enum.auto()
     strictly_decreasing = enum.auto()
     cumulative = enum.auto()
+    table = enum.auto()
     float_used = enum.auto()
 
 
 FLAG_TO_SRC_PREFIX = {
     Flags.alldifferent: 'include "alldifferent.mzn";',
     Flags.alldifferent_except_0: 'include "alldifferent_except_0.mzn";',
+    Flags.alldifferent_except: 'include "alldifferent_except.mzn";',
     Flags.all_equal: 'include "all_equal.mzn";',
     Flags.nvalue: 'include "nvalue_fn.mzn";',
     Flags.circuit: 'include "circuit.mzn";',
@@ -31,6 +34,7 @@ FLAG_TO_SRC_PREFIX = {
     Flags.decreasing: 'include "decreasing.mzn";',
     Flags.strictly_decreasing: 'include "strictly_decreasing.mzn";',
     Flags.cumulative: 'include "cumulative.mzn";',
+    Flags.table: 'include "table.mzn";',
 }